Created Kruskal's Algorithm Optimized in java
The code begins by defining a Edge class that represents an edge in the graph. It implements the Comparable interface to allow sorting based on the edge weights.

The DisjointSet class represents the disjoint-set data structure used to keep track of the disjoint sets during the algorithm. It implements the find and union operations.

The kruskalsAlgorithm method takes a list of edges and the number of vertices in the graph as input. It sorts the edges in non-decreasing order of their weights and initializes a DisjointSet object.

It then iterates through the sorted edges, checking if the source and destination vertices of each edge belong to different sets. If they do, the edge is added to the minimum spanning tree, and the sets are merged using the union operation.

In the main method, a sample graph with four vertices and edges is defined. The kruskalsAlgorithm method is called with the edges and number of vertices. The resulting minimum spanning tree is printed to the console.

You can modify the edges and numVertices variables to test the algorithm with different inputs.

#hacktoberfest2023